operacoes com marcadores (labels)

+++ bookmarks.py
@@ -1,3 +1,6 @@
+from tornado.database import Row
+import itertools
+
 class Bookmarks(object):
     
     def __init__(self, conn=None):
@@ -5,6 +8,7 @@ class Bookmarks(object):
            receive connection from Tornado Application
         """
         self.conn = conn
+        self.labels = Labels(conn)
 
     def insert(self, bookmark):
         insert_stmt = """
@@ -12,14 +16,53 @@ class Bookmarks(object):
                             bookmarks (name, description, url, user_id, published, updated) 
                         VALUES (%s, %s, %s, 1, UTC_TIMESTAMP(), UTC_TIMESTAMP())
                       """
-        return self.conn.execute(insert_stmt, bookmark['name'], bookmark['description'], bookmark['url'])
+        
+        bookmark_rowid = self.conn.execute(insert_stmt, bookmark['name'], bookmark['description'], bookmark['url'])
+        
+        if bookmark['labels']:
+            for label in bookmark['labels'].split(","):
+                label = self.labels.get_or_insert(label)
+                self.associate_label(bookmark_rowid, label['id'])
+
+        return bookmark_rowid
 
     def all(self):
         """
             retrieve all bookmarks
             TODO: paginate this query
         """    
-        return self.conn.query("SELECT * FROM bookmarks ORDER BY published DESC")
+        
+        query = """
+                    SELECT b.*
+                          ,l.id as label_id
+                          ,l.name as label_name 
+                    FROM bookmarks b 
+                         LEFT JOIN bookmarks_labels bl 
+                             on b.id = bl.bookmark_id 
+                         LEFT JOIN labels l 
+                                 on l.id = bl.label_id 
+                    ORDER BY b.published DESC
+                """
+        
+        results = self.conn.query(query)
+        
+        bookmarks_hash = {}
+        bookmarks = []
+        
+        
+        for result in results:
+            if bookmarks_hash.has_key(result['id']):
+                bookmarks_hash[result['id']].labels.append( Row(itertools.izip( ['id', 'name'], [ result['label_id'], result['label_name'] ] )) )
+            else:
+                bookmark = Row(itertools.izip(['id', 'name', 'url', 'description', 'published', 'labels']
+                                             ,[result['id'], result['name'], result['url'], result['description'], result['published'], [] ]))
+                if result['label_id']:
+                    bookmark.labels.append( Row(itertools.izip( ['id', 'name'], [ result['label_id'], result['label_name'] ] )) )
+                
+                bookmarks.append(bookmark)
+                bookmarks_hash[result['id']] = bookmark
+        
+        return bookmarks
 
     def delete(self, bookmark_id):
         """
@@ -28,6 +71,7 @@ class Bookmarks(object):
         bookmark = self.get_by_id(bookmark_id)
         if not bookmark: raise RecordNotFound
         
+        self.conn.execute("DELETE FROM bookmarks_labels WHERE bookmark_id = (%s)", int(bookmark_id))
         return self.conn.execute("DELETE FROM bookmarks WHERE id = (%s)", int(bookmark_id))
 
     def get_by_id(self, bookmark_id):
@@ -36,6 +80,28 @@ class Bookmarks(object):
         """
         return self.conn.get("SELECT * FROM bookmarks WHERE id = %s", int(bookmark_id))
 
+    def validate(self, bookmark):
+        messages = []
+        if not bookmark['url']:
+            messages.append("sounds good provide a url here...")
+            
+        return messages
+    
+    def is_associated(self, bookmark_id, label_id):
+        return self.conn.get("SELECT * FROM bookmarks_labels WHERE bookmark_id = %s AND label_id = %s", int(bookmark_id), int(label_id))
+    
+    def associate_label(self, bookmark_id, label_id):
+        associated = self.is_associated(bookmark_id, label_id)
+        if not associated:
+            self.conn.execute(
+                       "INSERT INTO bookmarks_labels (bookmark_id, label_id) VALUES (%s, %s)", bookmark_id, label_id)
+    
+    def disassociate_label(self, bookmark_id, label_id):
+        associated = self.is_associated(bookmark_id, label_id)
+        if associated:
+            self.conn.execute(
+                       "DELETE FROM bookmarks_labels WHERE bookmark_id = %s AND label_id = %s", bookmark_id, label_id)
+                
 class Labels(object):
 
     def __init__(self, conn=None):
@@ -48,7 +114,11 @@ class Labels(object):
         """
             retrieve a label or insert one if there is no label for name informed
         """	
-        label = self.conn.get("SELECT * FROM labels WHERE name = %s", label_name)
+        assert label_name
+        
+        label_name = label_name.strip()
+        
+        label = self.conn.get("SELECT * FROM labels WHERE LOWER(name) = %s", label_name.lower())
         
         if not label:  		      
             rowid = self.conn.execute(
@@ -79,7 +149,22 @@ class Labels(object):
         """
             remove a label by id
         """	
+        label = self.get_by_id(label_id)
+        if not label: raise RecordNotFound
+
+        self.conn.execute("DELETE FROM bookmarks_labels WHERE label_id = (%s)", int(label_id))
         return self.conn.execute("DELETE FROM labels WHERE id = (%s)", int(label_id))
     
 class RecordNotFound(Exception):
     pass
+
+class CustomRow(dict):
+    """
+        A dict that allows for object-like property access syntax.
+        copied from tornado.database.Row
+    """
+    def __getattr__(self, name):
+        try:
+            return self[name]
+        except KeyError:
+            raise AttributeError(name)
